Restructured code, added admin as app with subcommands.
