+++ dbtwiz/commands/__init__.py
@@ -1,0 +1,7 @@
+from .backfill import backfill
+from .build import build
+from .freshness import freshness
+from .model import model
+from .precommit import sqlfix
+from .test import test
+#from .example import example
